Solucion a que no me abria

+++ app/src/main/java/com/ucne/proj_1erparcial_ap2/ui/theme/Prestamos/RegistraPantallaViewModel.kt
@@ -46,7 +46,7 @@ class RegistroViewModel @Inject constructor(
     fun insertar() {
         val prestamos = PrestamosEntity(
 
-            deudor = deudor,
+            deudor = deudor.toInt(),
             concepto = concepto,
             monto = monto .toDoubleOrNull() ?: 0.0
         )
